feat: create rstack
