+++ packages/create-rsbuild/template-eslint/vanilla-js/eslint.config.mjs
