+++ packages/create-rsbuild/template-eslint/vanilla-js/package.json
