+++ packages/create-rsbuild/template-eslint/vanilla-ts/eslint.config.mjs
