+++ packages/create-rsbuild/template-eslint/vanilla-ts/package.json
